backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
click-default-group==1.2.4
colorama==0.4.6
fastapi==0.117.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.48.0
tabulate==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
pydantic[email]==2.11.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.3
httpx==0.25.2
